+++ src/main/CanDerivedStats.cpp
@@ -48,10 +48,10 @@ void CanDerivedStats::update(CanDiagnostics &diagnostics) noexcept {
                        .count();
 
   if (time_diff > 0) {
-    const auto update_rate = [&](std::deque<double> &window,
+    const auto update_rate = [&](std::deque<float> &window,
                                  std::optional<uint32_t> &value,
                                  uint32_t &last_value, int64_t time_diff) {
-      double rate = static_cast<double>(value.value() - last_value) / time_diff;
+      float rate = static_cast<float>(value.value() - last_value) / time_diff;
       window.push_front(rate);
       if (window.size() > window_size) {
         window.pop_back();
@@ -64,8 +64,8 @@ void CanDerivedStats::update(CanDiagnostics &diagnostics) noexcept {
 
     m_last_update = current_time;
 
-    const auto update_per_second = [&](const std::deque<double> &window,
-                                       std::optional<double> &per_second) {
+    const auto update_per_second = [&](const std::deque<float> &window,
+                                       std::optional<float> &per_second) {
       if (!window.empty()) {
         per_second =
             std::accumulate(window.begin(), window.end(), 0.0) / window.size();
